utils/requirements.txt
flask
gunicorn
tensorflow-cpu
numpy
pillow
matplotlib
reportlab
